feat: Add context requirements for actions in state transitions (WEB-3866)
An update has been made to the tests/RequiredContextTest.php, src/Definition/MachineDefinition.php and a new file, tests/Stubs/Actions/IsOddAction.php, has been created. The changes include requiring context values for both guards and actions in state transitions in the Event Machine. A new test was added to validate this concept while MachineDefinition was updated to throw an exception if a required context object is missing. Lastly, a new stub class, IsOddAction, was created to test this new feature.