(WEB-3874) refactor: context validation in behaviors
Refactored the context validation logic in TransitionDefinition and MachineDefinition classes. Instead of using a static method to check for missing context and raising an exception, moved the validation responsibility to the behaviors themselves (GuardBehavior and InvokableBehavior). It gives more granular control over context validation and decreases the coupling between classes. Code is more object-oriented now.