(WEB-3893) feat: Add getType method to InvokableBehavior
This commit adds a getType method to the InvokableBehavior class. It returns the type of the behavior in string, which is determined by converting the fully qualified class name (FQCN) to the base class name in camel case. This feature taps into the utility of the Illuminate\Support\Str class to make this conversion.