(WEB-3901) refactor: Refactor machine event creation for history tracking
This change rectifies the way 'root_event_id' is assigned in the creation of new MachineEvent instances for maintaining history. Earlier, for the first event the 'root_event_id' was being set as its own id. Now, it is set to the id of the first event in history, else (if no events exist in history), it is set to its own id. This adjustment promotes more accurate and consistent event tracking.