(WEB-3901) refactor: Remove deprecated usage of Str class for event name formatting
Revised the method of generating internal event names within the 'Actor/State.php' file. Removed the dependency on the Illuminate\Support\Str class and instead used the generateInternalEventName method from the type object. This change improves the readability and maintainability of the code, and also allows the inclusion of machineId in the event name.