(WEB-3901) feat: Update guard behavior in TransitionDefinition
Refactored the guard behavior in TransitionDefinition to improve code readability. Previously, the generation of event name was dependent on the deprecated usage of Str class. The changes switched to using 'generateInternalEventName' from the type object. This not only enhances code readability but also ensures maintainability while allowing the inclusion of machineId in the event name.