(WEB-3900) test: Update event status from INIT to ENTER
The code was modified in several test files (EventMachineTest.php, RaisedEventsTest.php, EventStoreTest.php) to update the status of an event from 'state.<name>.init' to 'state.<name>.enter'. This change was done to make test scenarios more aligned with the actual sequence of steps in state machine events, where 'enter' is closer to real-world semantic than the previous 'init'.