(WEB-3900) refactor: Rename ACTION_INIT to ACTION_START in event definitions
Resolved inconsistencies in the naming of the event action. Instead of using "init" in the event name, we are now using "start". The change was made in src/Definition/InternalEvent.php and src/Definition/MachineDefinition.php and the corresponding changes have been reflected in the test files: EventStoreTest.php and RaisedEventsTest.php. This change aligns the language used across the codebase to denote the start of an action.