(WEB-3900) Add STATE_EXIT event to MachineDefinition
Added a STATE_EXIT internal event in the MachineDefinition file. This event is recorded just before transitioning to a new state. The 'xyz.state.{stateKey}.exit' pattern was also added to RaisedEventsTest and verified in the EventStoreTest, increasing the count of expected events. This change was made to keep track of when and where state transitions happen, increasing observability.