(WEB-3900) refactor(MachineDefinition@transition): Refactor code comments for better understanding
Code comments were updated to be more straightforward and explanatory. Changes were made to remove ambiguity and enhance readability. Expanded some previously shortened phrases and also improved phrasing for clarity. An example is that, instead of saying "Find the transition definition for the event type", we now have "Get the transition definition for the current event type." Clear and comprehendible comments can greatly assist other developers in understanding the functionality of the code.