(WEB-3900) chore: Update comment for transition start event
The comment for the event recording block within the MachineDefinition function has been revised. Previously, it was mistakenly referring to a "state entry start event". However, this section of code is actually handling the beginning of a transition, so it has been updated to reflect that with "transition start event" for better accuracy and clarity in understanding the code functionality.