(WEB-3900) feat: Reorder execution of transition actions and exit actions
The execution order of transition actions and exit actions was modified in the 'MachineDefinition.php'. Now, the transition actions are executed before the exit actions. This change helps to ensure that any state changes caused by transition actions are taken into account before running the exit actions on a given state.