(WEB-3900) feat: Add transition start event and placeholder in MachineDefinition
In the process of running transitions in MachineDefinition, it's necessary to record when transitions start for future debugging and data tracking. A new block has been added to set an internal event representing the beginning of a transition, also specifying the state and type of event behavior as a placeholder for context. A TODO comment was also included indicating that further failure handling is needed if no valid transition branch is found.