(WEB-3900) test: Add transition finish events to tests
Two tests - RaisedEventsTest and EventStoreTest - have been updated to include transition finish events. This change ensures that we are properly verifying the behaviour of transitions, checking not only their initiation but also their conclusion. This adds thoroughness to our testing and ensures we cover all event branches in the system.