(WEB-3903) feat: Add exception handling for invalid guarded transitions
Extended the guard initialization to handle InvalidGuardedTransitionException. When a transition is guarded, the added code checks for class existence and if the guard is a subclass of ValidationGuardBehavior, throwing an exception if true. This is to address the scenario when a validation guard is used inside a guarded transition which is not allowed, hence ensuring our Guard adheres to the Single Responsibility Principle. Testing for this scenario has also been added.