(WEB-3904) feat: Add state definition type in StateDefinition class
A Type property (`$this->type`) was added to the StateDefinition class. It is assigned by calling the newly added method `getStateDefinitionType()`. This method returns the type of the state, primarily to handle different functionalities for 'final' state. This feature was added to implement rules specific to different state types, such as 'final' state, which according to further tasks cannot have child states or transitions.