(WEB-3904) refactor: Improves error message clarity in final state exceptions
This commit modifies the error message in the `InvalidFinalStateDefinitionException` to be more descriptive and actionable. Rather than simply stating that the final state shouldn't have child states, the new message provides additional instructions asking to revise the state machine definitions to ensure correct configuration of final states. This change is reflected in the `StateDefinitionTypeTest` where the expected exception message is updated accordingly.