(WEB-3900) chore: Remove unused Exception and refine doc-blocks
The 'BehaviorNotFoundException' exception import was removed as it was not being used. This reduces unnecessary clutter in the code. Furthermore, full namespace paths for exceptions are now provided in the constructor doc-block, improving readability and ensuring correct reference if the same-named exceptions are used from a different namespace in the future.