(WEB-3900) feat: Add machine finish event for initial final state
Enhanced syntax in MachineDefinition by incorporating a check for the initial state to be a final state, in which case, an internal event behavior for MACHINE_FINISH is recorded. This impacts the state machine where an initial state can now directly jump to a final state.
Adjusted relevant unit test to account for this new functionality in StateDefinitionTypeTest. The 'machine.finish' event is now expected in the event history for an initial final state.