(WEB-3900) feat: Add machine finish event handling for final states
Introduced handling for cases where a state machine's initial state is set as a final state. The implemented logic records a machine finish event in such scenarios. Additionally, this diff introduces a new Exception which is thrown if the machine context is found missing when getting the initial state. Corresponding tests were included to ensure the proper functioning of these changes.