(WEB-3905) refactor: Change base class for NoTransitionDefinitionFoundException
The diff shows that the base class for NoTransitionDefinitionFoundException was changed from RuntimeException to LogicException. The RuntimeException is replaced by LogicException as it is a more suitable choice given the nature of this specific event - it is thrown when the program encounters an error in its logic. This change enhances the clarity and meaningfulness of the exception in accordance with the Standard PHP Library (SPL) hierarchy.