(WEB-3905) chore: Add docstring to BehaviorType class
In order to improve comprehension and usage of code for our new developers, a docstring providing a detailed explanation has been added to the BehaviorType class within the Behavior subsystem. This new commenting clearly defines its purpose and function, enhancing the readability and contributing towards the maintainability of the codebase.