(WEB-3905) chore: Add doc comments to RestoringStateException class
In order to provide a better understanding of the purpose and usage of RestoringStateException class, doc comments have been added. This detail includes explaining what the class represents and its extending to RuntimeException class. A comment has also been added to the build function within the class describing its purpose and return. This improves code readability and future debugging efforts.