(WEB-3924) feat: Refactor EventMachine and MachineActor classes
Deleted `EventMachine` class as it was no longer necessary. Introduced new functionalities in `MachineActor` class, which is also renamed to `Machine`.

The `EventMachine` class was deleted because its functionalities were moved to `MachineActor`. This resulted in a more cohesive design, with the `MachineActor` handling both the machine definition and event processing, as well as starting the machine as an actor. The `MachineActor` class was also renamed to `Machine` to better reflect its enhanced functionalities. Detailed comments were added for each method to guide future developers working with this code.