(WEB-3924) test: Change class MachineActor to Machine across various test files
This commit includes changes across various test files to replace instances of the class "MachineActor" with "Machine". These changes are necessary due to the refactoring of the "MachineActor" class in the core code. The adjustments ensure that the tests are compatible with the updated class structure and function as intended. The replacements were made although no changes in functionality are introduced by this commit. In all test instances, the class "MachineActor" was simply swapped with the class "Machine".