(WEB-3936) test: Refactor code to support TrafficLightsContext in action functions
The '__invoke' methods in the 'IncrementAction', 'DecrementAction', 'MultiplyByTwoAction', 'AddAnotherValueAction', 'SubtractValueAction', and 'AddValueAction' classes were updated to support a 'TrafficLightsContext' or 'ContextManager' as their first parameter. This change will allow greater control over the context in which these methods are executed. The guard class 'IsEvenGuard' was also updated to accept these types for the context. These changes align the type specifications of these actions and guards for better code readability and maintainability.