refactor: Modified fetching machine definition logic in Machine.php
Changed the mechanism of fetching the machine's definition in the class 'Machine.php'. Previously, the definition() method was returning null, now it throws a 'MachineDefinitionNotFoundException' if the definition is not found. The purpose of this change is to handle the exception promptly and provide a clearer error message when the machine definition is not found. The initiation of machine state has also been streamlined to improve the code readability.