Add: Create MachineFacade class for Event Machine
Implemented a new 'MachineFacade' class under the namespace 'Tarfinlabs\EventMachine\Facades' to access the event machine using the facade pattern. The purpose of this addition is to provide a simple and stable interface to access the 'event-machine' definition, thus optimizing the process and enhancing code maintainability. This class includes a getFacadeAccessor method that returns the 'event-machine'.