(WEB-3973) Remove unused exception class and its reference
Deleted the InvalidGuardedTransitionException class and removed its reference from the StateDefinition class. The exception was designed to prevent usage of ValidationGuardBehavior inside guarded transitions. However, in the current implementation, no guard uses this specific behavior, so keeping this exception was unnecessary and was complicating the code unnecessarily. This made the code simpler and leaner.