(WEB-3973) Removed unused exception class and its reference
The InvalidGuardedTransitionException class and its reference from GuardedActionsTest.php have been removed. This exception was initially introduced to catch the use-case of ValidationGuardBehavior in guarded transitions. However, it was not used anywhere in the codes and was unnecessarily complicating the code structure. The code is significantly simplified with this removal as part of the ongoing refactoring task under ticket WEB-3973.