(WEB-3980) test: Correct test instantiation in StateDefinitionTypeTest
Modified the instantiation of Machine in the StateDefinitionTypeTest to pass 'definition' directly instead of under a 'config' key'. This was causing the tests to fail as the Machine constructor expects 'definition' to be passed directly. This change was made to ensure the proper functioning of unit tests by adhering to the expected constructor parameters.