Allow string events in Machine's send method
Modified the 'send' function in Machine class to allow events to be passed as strings. This change facilitates easier event handling by assigning the string directly to the 'type' attribute, if an event is passed as a string. This enhances flexibility and simplifies the process of event state transition, contributing to more efficient code execution.