feat: Add getMorphClass method to ContextManager
This commit adds a new public method, getMorphClass, to the ContextManager class. This method was added to provide a way to easily retrieve the name of the controller's associated model. It is particularly useful in polymorphic relationships, where you might need to know the class name of the associated model at runtime.