(WEB-3985) feat: Add function to return machine's final state result
In the Actor/Machine.php, a new function called 'result' was added. This function checks if the machine's current state is final, and if so, returns a result based on the current state's behavior and context. This was necessary to allow the retrieval of a final result when the machine's state has reached its end.