(WEB-3985) feat: Add abstract ResultBehavior class
Added an abstract ResultBehavior class in the Behavior directory for the EventMachine namespace. This new class, as an extension of the InvokableBehavior class, introduces a new behavior that the system can use. The method is meant to be invoked with a ContextManager, EventBehavior, and a set of arguments. The specific purpose of this class should be defined in future commits or implementations.