(WEB-3985) feat: Add result initialization to StateDefinition
Added the 'initializeResults' function to 'StateDefinition' to handle result set up in final state type. Prior to this, there was no handling for results in the state machine for the final state, impacting functionality required for result processing. With this change, results are initialized in the machine's behavior if a final state is detected and a result is set in the configuration.