(WEB-3985) refactor: Refactor result() function in Machine.php for accuracy
The 'result()' function in 'Machine.php' was refactored to improve the state machine execution's reliability. This change improves the accuracy by correctly applying the associated result behavior in the machine's final state, taking into consideration the 'currentStateDefinition'. Previously, if the 'currentStateDefinition' was isn't set correctly, the function could potentially lead to undesirable results. Resolution of this issue ultimately enhances the function's dependability.