(WEB-3993) refactor: Change state path from array to string
In various instances of code, `currentStateDefinition->path` is replaced by `currentStateDefinition->route`. Previously, `currentStateDefinition->path` returned an array which was then converted to a string by calling the `implode` function each time it was used. This resulted in the overuse of the `implode` function.

To simplify this, the state path was changed from array to string, and `currentStateDefinition->route` now directly returns a string. This reduces superfluous function calls and enhances the readability and maintainability of the code.