(WEB-3991) feat: Update initialState handling in StateDefinition class
The previous version assumed a null value when there was no initialStateDefinitionKey. However, this version considers the root state definition to be the initial state if there is no initialStateDefinitionKey. It also introduced several comments to explain this process, improving readability for future code maintenance. Finally, it adds recursive calls for initial state definition with children, enhancing the flexibility and robustness of the state machine structure.