(WEB-3991) test: Add test to check transition between nested and top-level states
This commit aims to ensure correct transition between states in the event machine. A new use instruction to include ContextManager was added. An extra case to test 'Top-Level Transitions' was added to 'ForbiddenTransitionTest'. This new test covers a scenario where the machine transitions from a nested state to a top-level state, which increases the test coverage and guarantees the robustness of the code.