Update: Enable test coverage in Github actions workflow
The "coverage" field of the "Setup PHP" action has been set to "pcov", and the "Execute tests" command has now a "--coverage" option. This is done to generate a coverage report for the test cases in the Pest test suite. PCOV has been chosen as it is a faster and lightweight code coverage driver.