Adjust test execution order in Github workflow file
The "--coverage" option during test execution has been moved up in the ordering to ensure that it takes precedence. This is important because we want to make sure the code coverage report is generated even if there are issues with any of the following options. The function of other options remains the same, with "--colors" set to always display test outcome colorfully and "--order-by" to random for avoiding sequential biases in testing.