Add automated testing workflow with Pest
A new pest-coverage.yml file has been introduced to the .github/workflows directory to automate the execution of tests with Pest. This testing will run on every push or pull request made to the 'main' branch. This workflow has been added to ensure maintainability of high-quality code and encourage a standard practice of having 100% code coverage with every update.