refactor: Update how `runAction` method handles `InvokableBehavior` in `MachineDefinition.php` (WEB-4344)
This commit modifies the `runAction` method in `src/Definition/MachineDefinition.php`. The way `InvokableBehavior` instances are handled has been updated. Instead of using a `ReflectionMethod` to get the parameters of the `__invoke` method, a `ReflectionFunction` is now used on the result of the `definition()` method. The action behavior is then executed by invoking the `definition()` method directly if the behavior is an instance of `InvokableBehavior`, or the behavior itself otherwise.